conversion to ONNX sometimes depends on the runtime

+++ _unittests/ut_onnxrt/test_onnxrt_validate_type.py
@@ -56,7 +56,7 @@ class TestOnnxrtValidateType(ExtTestCase):
                 enumerate_validated_operator_opsets(
                     verbose, debug=True, fLOG=myfLOG, dump_folder=temp,
                     models=models, filter_exp=filter_exp,
-                    dtype=dtype, opset_min=11, store_models=True)):
+                    opset_min=11, store_models=True)):
 
             up = {}
             outputs = []
